= fixed, hello, keepalive, notify, join as per nunki's chool

+++ iniParser.h
@@ -1,6 +1,24 @@
-#include<stdio.h>
-#include<iostream>
-#include<list>
+#include <errno.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <iostream>
+#include <sys/types.h>
+#include <sys/socket.h>
+#include <netdb.h>
+#include <netinet/in.h>
+#include <arpa/inet.h>
+#include <sys/wait.h>
+#include <signal.h>
+#include <list>
+#include <unistd.h>
+#include <ctype.h>
+#include <map>
+#include <openssl/sha.h>
+#include <stdlib.h>
+#include <string.h>
+#include <set>
+
 
 using namespace std;
 
@@ -18,7 +36,7 @@ struct myStartInfo
 	unsigned char homeDir[512];
 	unsigned char logFileName[256];
 	unsigned int autoShutDown;
-	unsigned int ttl;
+	uint8_t ttl;
 	unsigned int msgLifeTime;
 	unsigned int getMsgLifeTime;
 	unsigned int initNeighbor;
@@ -36,7 +54,7 @@ struct myStartInfo
 	bool isBeacon;						// to check if the current node is beacon or not
 	unsigned char node_id[265] ;					// To store the unique node ID of this node
 	unsigned char node_instance_id[300];				// To store the node instance ID, node_id + time when node started
-	unsigned int status_ttl ;
+	uint8_t status_ttl ;
 	unsigned char status_file[256] ;
 	unsigned int statusResponseTimeout ;
 	unsigned int checkResponseTimeout ;
